+++ tests/codegen/paginator-tests/src/main/kotlin/com/test/endpoints/EndpointParameters.kt
@@ -1,0 +1,3 @@
+package com.test.endpoints
+
+class EndpointParameters