+++ tests/codegen/paginator-tests/src/main/kotlin/com/test/endpoints/EndpointProvider.kt
@@ -3,4 +3,4 @@ package com.test.endpoints
 /**
  * Stubbed EndpointProvider since we don't use a concrete protocol generator for this test.
  */
-typealias EndpointProvider = aws.smithy.kotlin.runtime.client.endpoints.EndpointProvider<Unit>
+typealias EndpointProvider = aws.smithy.kotlin.runtime.client.endpoints.EndpointProvider<EndpointParameters>